Finalised Ship movement (for now)
Modified the previous system to use a custom camera (rather than cinemachine) to provide more flexibility especially given the nature of the game and how the camera needs to manipulate around a planet seamlessly. This required a rework of the movement logic and the deprecation of the CameraMovement script however the current system is a lot more simple and efficient.