Additional spawn logic
I previously started working on a system to spawn and despawn enemies when the player was in or out of distance (save computation). I modified this logic to not destroy enemies attacking the player, they retreat when too far from base and the base carries info on how many enemies were despawned as to not over spawn.