Fixed bug, improved code design
Enemies would sometimes not chase the player even when they were in range because a raycast would check if there was an obstacle in the way, This would hit other enemies creating the issue. I also moved the layer mask references into variables accessible from the editor and available to the whole class. This reduces issues if there happens to be things like renaming variables.