Modified enemy ai logic for teamwork
Ai now will notify the main base which will then notify all the other ai. I also added a check into the on trigger enter method. This makes sure we only perform checks and operations if we aren't already attacking. This reduces wasted computation and also the main purpose which was to make sure every enemy isn't trying to alert the base and all trying to get component at the same time.